#21 マイページ以下の遷移構造を変更
